% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alligatorLength.R
\docType{data}
\name{alligatorLength}
\alias{alligatorLength}
\title{Study of Diets in Alligators at Lake George, Florida}
\format{
The data frame has 63 rows and 3 variables:
\describe{
 \item{sex}{Female (F) or Male (M).}
\item{length}{Length of alligator in meters. Subadult alligators have length < 1.83 and adults if > 1.83 meters. }
\item{foodChoice}{Primary stomach contents of the alligator.}
}
}
\usage{
data(alligatorLength)
}
\description{
Data.frame
}
\details{
A study in Lake George, Florida caught 63 alligators.
Each alligator's stomach contents were classified as fish, invertebrate, or other.
The sex and the length of the alligator were also recorded.
}
\examples{
data("alligatorLength", package='syllogi')
str(alligatorLength)
}
\references{
Agresti, A. (2013) Categorical Data Analysis. 3rd Edition, John Wiley & Sons, Hoboken, New Jersey.
}
\keyword{data}
