% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualSales.R
\docType{data}
\name{annualSales}
\alias{annualSales}
\title{Fictitious Data Set of Annual Sales}
\format{
The data frame has 12 rows and 3 variables:
\describe{
  \item{sales}{Annual gross sales in $1000 of dollars.}
  \item{advert}{Annual cost of advertising in $1000 of dollars.}
\item{quality}{Quality of their store\'s typical product: 0=very poor quality to
25 = exceptional quality.}
}
}
\usage{
data(annualSales)
}
\description{
Data.frame
}
\details{
You are hired as a statistical consultant.
Twelve stores in the Fort Collins, CO area have asked you to develop a prediction model for their annual gross sales (sales; measured in $1000 of dollars).
They would like to know if it is possible to predict the amount of their sales
by knowing how much they spend annually on advertising
(advert; measured in $1000 of dollars)
and the quality of their store’s typical product
(quality; measure on a scale from 0 = very poor quality to 25 = exceptional quality).
}
\examples{
data("annualSales", package='syllogi')
str(annualSales)
}
\references{
fictitious data set
}
\keyword{data}
