% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beer.R
\docType{data}
\name{beer}
\alias{beer}
\title{Beer}
\format{
The data frame has 86 rows and 5 variables:
\describe{
\item{brand}{Brand name of the beverage}
\item{brewery}{Brewery of the beverage}
\item{percentAlcohol}{Percent alcohol by volume}
\item{calories}{Total calories }
\item{carbohydrates}{Total carbohdrates} 
 }
}
\usage{
data(beer)
}
\description{
Data.frame
}
\details{
Does a beer with more carbohydrates tend to have more alcohol?
To answer this question the carbohydrates and percent alcohol from several different beer brands were measured.
}
\examples{
data("beer", package='syllogi')
str(beer)
}
\keyword{data}
