% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butterflyPlot.R
\docType{data}
\name{butterflyPlot}
\alias{butterflyPlot}
\title{Fictitious Data Set of Butterfly Counts}
\format{
The data frame has 40 rows and 2 variables:
\describe{
  \item{area}{Plot area size in hectares.}
  \item{numSpecies}{Count of number of unique species.}
}
}
\usage{
data(butterflyPlot)
}
\description{
Data.frame
}
\details{
Plots ranging in size from 1ha to 1000ha, were left uncut in a larger landscape of logged tropical rainforest.
  In each  plot the number of unique butterfly species was recorded.
  What is  the relationship between plot size and unique species count?
}
\examples{
data("butterflyPlot", package='syllogi')
str(butterflyPlot)
}
\references{
fictitious data set
}
\keyword{data}
