% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogFood.R
\docType{data}
\name{dogFood}
\alias{dogFood}
\title{Fictitious Data Set Comparing Dog Food Brands}
\format{
The data frame has 25 rows and 2 variables:
\describe{
  \item{type}{The type of dog food: our dog food or one of the four top sellers.}
  \item{gain}{The percent weight gain.}
}
}
\usage{
data(dogFood)
}
\description{
Data.frame
}
\details{
You are hired as a statistical consultant for a dog food manufacturing company.
The engineers who designed the company's dog food would like to know how it compares to the current top selling dog food brands on the market?
To answer this question, 25 puppies of the same breed and age (within a week of each other) were chosen for this study.
Five puppies were assigned to each dog food type.
After 4 weeks the percent of weight gained for each puppy was determined.
}
\examples{
data("dogFood", package='syllogi')
str(dogFood)
}
\references{
fictitious data set
}
\keyword{data}
