% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/federalistPapers.R
\docType{data}
\name{federalistPapers}
\alias{federalistPapers}
\title{Word count of "may" in Federalist Papers}
\format{
The data frame has 509 rows (a block of words) and 2 variables:
\describe{
  \item{author}{Hamilton or Madison.}
  \item{wordCount}{Count of the word "may" within that block.} 
}
}
\usage{
data(federalistPapers)
}
\description{
Data.frame
}
\details{
Mosterller and Wallace (1964) recorded the frequency of use of various words in selected articles by Alexander Hamilton and James Madison.
The articles were divided in blocks of about 200 words.
They recorded the count of various words.
Provided is the count for the word 'may'.
}
\examples{
data("federalistPapers", package='syllogi')
str(federalistPapers)
}
\references{
Mosteller, F. and D. L. Wallace. Inference and Disputed Authorship: The Federalist. Reading, MA., 1964

Gelman, A., Carlin, J.B., Stern, H.S., Dunson, D.B., Vehtari, A., & Rubin, D.B. (2013). Bayesian Data Analysis (3rd ed.). Chapman and Hall/CRC. https://doi.org/10.1201/b16018
}
\keyword{data}
