% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frogs.R
\docType{data}
\name{frogs}
\alias{frogs}
\title{Frog Presence}
\format{
The data frame has 212 rows and 6 variables:
\describe{
  \item{present}{1 = frogs are present, 0 = not.}
\item{distance}{Distance to water in dekameters.}
\item{meanmin}{Average spring minimum temperate in centigrade.}
}
}
\usage{
data(frogs)
}
\description{
Data.frame
}
\details{
A biologist conducted a study to determine the presence of frogs.
Random locations were selected within the study area.
}
\examples{
data("frogs", package='syllogi')
str(frogs)
}
\references{
Unknown
}
\keyword{data}
