% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericData.R
\docType{data}
\name{genericData}
\alias{genericData}
\title{Generic Data Set}
\format{
An object of class \code{data.frame} with 60 rows and 7 columns.
}
\usage{
data(depression)
}
\description{
Generic data set with four ratio predictors (X1,X2,X3,X4), two categorical predictors (A,B) and one ratio response variable (Y).
}
\details{
This is a fictious dataset useful for teaching how to use and interpret linear statistical models.
}
\references{
fictitious data set
}
\keyword{data}
