% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golf.R
\docType{data}
\name{golf}
\alias{golf}
\title{Golfing}
\format{
The data frame has 18 rows and 3 variables:
\describe{
\item{clubs}{clubs used for that round of golf}
\item{course}{course for the round of golf}
\item{score}{score or strokes for 18 holes}
 }
}
\usage{
data(golf)
}
\description{
Data.frame
}
\details{
I purchased new golf clubs last summer, which I believe will significantly improve my game.
I recorded my score after three rounds of golf with my new clubs and my old clubs.
I also played at three different courses.
}
\examples{
data("golf", package='syllogi')
str(golf)
}
\references{
fictitious data set
}
\keyword{data}
