% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homicide.R
\docType{data}
\name{homicide}
\alias{homicide}
\title{Detroit Homicides}
\format{
The data frame has 13 rows and 12 variables:
\describe{
  \item{homicide}{Number of homicides per 100k population.}
  \item{police}{Number of full-time police officers per 100k population.}
\item{unemp}{Percent unemployed in the population.}
\item{mfWork}{Number of manufacturing workers (thousands).}
\item{gunLic}{Number of handgun licences per 100k population.}
\item{gunReg}{Number of handgun registered per 100k population.}
\item{hArrest}{Percent of homicides cleared by arrests.}
\item{whiteMale}{Number of white males.}
\item{nmfWork}{Number of non-manufacturing workers (thousands).}
\item{govWork}{Number of government workers (thousands).}
\item{hourEarn}{Average hourly earnings (dollars).}
\item{weekEarn}{Average weekly earnings (dollars).}
}
}
\usage{
data(homicide)
}
\description{
Data.frame
}
\details{
Homicides per capita in Detroit from years 1961-1973. Several other metrics on Detroit are also included.
}
\examples{
data("homicide", package='syllogi')
str(homicide)
}
\references{
Unknown
}
\keyword{data}
