% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutritionCancer.R
\docType{data}
\name{nutritionCancer}
\alias{nutritionCancer}
\title{Nutrition Cancer Study}
\format{
The data frame has 50 rows and 6 variables:
\describe{
  \item{id}{Patient's ID number.}
\item{age}{The age of the patient in years.}
\item{length}{The duration or time in months the patient has had breast cancer.}
\item{serving}{The number of servings the patient eats of fruits and vegetables in a typical day.}
\item{familyHistory}{Does or did any blood relatives (i.e. mother, grandmother, aunt, etc.) have or had breast cancer?}
  \item{stage}{The stage of the cancer: 0-non-invasive to IV-very invasive or “advanced” cancer.}
}
}
\usage{
data(nutritionCancer)
}
\description{
Data.frame
}
\details{
Fictitious data set for teaching purposes. The fictitious scenario:

The purpose of a medical study is to examine the relationship between eating fruits and vegetables and breast cancer.
To study the relationship, 1500 caucasian women with breast cancer were randomly selected from the list of cancer patients in the U.S.
The first 50 patients have been measured.
}
\examples{
data("nutritionCancer", package='syllogi')
str(nutritionCancer)
}
\references{
Fictitious data set
}
\keyword{data}
