% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patientSatisfaction.R
\docType{data}
\name{patientSatisfaction}
\alias{patientSatisfaction}
\title{Patient Satisfaction}
\format{
The data frame has 46 rows and 4 variables:
\describe{
  \item{satisfaction}{Patient's level of satisfaction, higher value means more satisfied.}
  \item{age}{Patient's age in years.}
\item{severityIllness}{Patient's severity of illness, higher value means more sever.}
\item{anxietyLevel}{Patient's anxiety level, higher value means more sever.} 
}
}
\usage{
data(patientSatisfaction)
}
\description{
Data.frame
}
\details{
A hospital administrator wants to predict patient's satisfaction using their
age, severity of illness, and anxiety level. Forty six patients were selected for the study.
}
\examples{
data("patientSatisfaction", package='syllogi')
str(patientSatisfaction)
}
\references{
Kutner, M. H., Nachtsheim, C., Neter, J., & Li, W. (2005). Applied linear statistical models (5th ed.). McGraw-Hill Irwin.
}
\keyword{data}
