% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politicalIdeology.R
\docType{data}
\name{politicalIdeology}
\alias{politicalIdeology}
\title{Political Ideology}
\format{
The data frame has 20 rows and 4 variables:
\describe{
 \item{gender}{Female (F) or Male (M).}
\item{party}{Democrat (D) or Republican (R)}
\item{ideol}{Very liberal (VL), Slightly Liberal (SL), Moderate (M),
Slightly conservative (SC), or Very conservative (VC).}
\item{count}{Count of people.}
}
}
\usage{
data(politicalIdeology)
}
\description{
Data.frame
}
\details{
A 1991 U.S. General Social survey that cross classifies people according to gender,
political party, and political ideology.
}
\examples{
data("politicalIdeology", package='syllogi')
str(politicalIdeology)
}
\references{
Bilder, C.R., & Loughin, T.M. (2014). Analysis of Categorical Data with R (1st ed.). Chapman and Hall/CRC. https://doi.org/10.1201/b17211
}
\keyword{data}
