% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolProgram.R
\docType{data}
\name{schoolProgram}
\alias{schoolProgram}
\title{High School and Beyond Survey}
\format{
The data frame has 200 rows (a student) and 11 variables:
\describe{
  \item{id}{Student ID.}
  \item{gender}{Student's gender.}
  \item{race}{Student's race.}
  \item{ses}{Socio economic status of the student's family, with levels low, middle, and high.}
  \item{schtype}{ Type of school: public or private.}
  \item{prog}{Type of program the student wants to attend after high school.}
  \item{read}{Student's standardized reading score.}
  \item{write}{Student's standardized writing score.}
  \item{math}{Student's standardized math score.}
  \item{science}{Student's standardized science score}
  \item{scost}{Student's standardized social studies score}
 }
}
\usage{
data(schoolProgram)
}
\description{
A survey conducted on high school seniors by the National Center of Education Statistics.
}
\details{
Two hundred students were randomly selected from the whole cohort in the survey.
}
\examples{
data("schoolProgram", package='syllogi')
str(schoolProgram)
}
\references{
https://www.openintro.org/data/index.php?data=hsb2

UCLA Institute for Digital Research & Education - Statistical Consulting.
}
\keyword{data}
