% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipDamage.R
\docType{data}
\name{shipDamage}
\alias{shipDamage}
\title{Wave Damage of Ships}
\format{
The data frame has 20 rows and 5 variables:
\describe{
\item{shipType}{Type of ship}
\item{constYear}{Year of construction}
\item{operation}{Period of operation}
\item{months}{Aggregate months of service}
\item{incidents}{Number of damage incidents}

 }
}
\usage{
data(shipDamage)
}
\description{
Data.frame
}
\details{
Cargo carrying vessel's damage to the forward section due to waves.
These data are only for the period of operation 1975 to 1979.
}
\examples{
data("shipDamage", package='syllogi')
str(shipDamage)
}
\references{
McCullagh, P. and Nelder, J.A. (1989) Generalized Linear Models. 2nd Edition, Chapman and Hall, London.
}
\keyword{data}
