% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipGold.R
\docType{data}
\name{shipGold}
\alias{shipGold}
\title{Ships and Gold}
\format{
The data frame has 20 rows (a ship) and 2 variables:
\describe{
  \item{shipSize}{Size of the ship measured in inches on the horizon.}
  \item{wordCount}{Amount of gold pieces on the ship.} 
}
}
\usage{
data(shipGold)
}
\description{
Data.frame
}
\details{
Fictitious data set for teaching purposes. The fictitious scenario:

Captain Buck Tooth has taken you prisoner aboard his pirate ship, the Lucky Lemon.
He sees from your college transcripts you have taken a couple of statistics courses.
Captain Buck Tooth wants you to predict the amount of gold a ship is carrying based on the size of the ship.
Specifically, he thinks bigger ships carry more gold.
For the last several ships he has looted he measured the height in inches when the ship was still way off on the horizon.
The captain also has a good memory and remembers how much gold was taken from each ship in number of pieces.
}
\examples{
data("shipGold", package='syllogi')
str(shipGold)
}
\references{
Fictitious data set
}
\keyword{data}
