% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soreThroat.R
\docType{data}
\name{soreThroat}
\alias{soreThroat}
\title{Sore Throat}
\format{
The data frame has 35 rows and 3 variables:
\describe{
  \item{duration}{Duration of the surgery in minutes.}
  \item{type}{Type of device used to secure the airway: laryngeal mask airway or tracheal tube.}
  \item{sore}{Does the patient experience a sore throat on waking: 0 = No, 1 = Yes.}
}
}
\usage{
data(soreThroat)
}
\description{
Data.frame
}
\details{
A study of patients having surgery with general anesthesia.
The research questions is does the patient experience a sore throat upon walking?
}
\examples{
data("soreThroat", package='syllogi')
str(soreThroat)
}
\references{
Agresti, A. (2002) Categorical Data Analysis. 2nd Edition, John Wiley & Sons, Hoboken, New Jersey.
}
\keyword{data}
