% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volDiamTree.R
\docType{data}
\name{volDiamTree}
\alias{volDiamTree}
\title{Tree Volumn and Diameter}
\format{
The data frame has 70 rows and 2 variables:
\describe{
  \item{diam}{Diameter of the tree in inches.}
  \item{vol}{Volume of the tree in cubic feet.}
}
}
\usage{
data(volDiamTree)
}
\description{
Data.frame
}
\details{
Many difference groups (lumber industry, ecologists, foresters, etc.) benefit from being able to predict the volume of a tree just by knowing its diameter.
The diameter and the volume of trees was recorded.
Bigger diameter trees have more volume but what is the exact relationship?
}
\examples{
data("volDiamTree", package='syllogi')
str(volDiamTree)
}
\references{
Unknown
}
\keyword{data}
