% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightLoss.R
\docType{data}
\name{weightLoss}
\alias{weightLoss}
\title{Weight Loss Study}
\format{
The data frame has 60 rows and 2 variables:
\describe{
  \item{drug}{Which weight loss drug the participant took for 6 weeks.}
  \item{loss}{Percent of weight loss after the 6 weeks. }
}
}
\usage{
data(weightLoss)
}
\description{
Data.frame
}
\details{
Fictitious data set for teaching purposes. The fictitious scenario:

 You are a statistical consultant.
A client comes to you asking for help with their analysis.
The client is from a drug company.
Their new drug is supposed to help people lose weight.
They conducted an experiment with their drug (drug A) and the two best selling weight loss drugs (B and C).
Male participants from age 50-60 were used in the study.
Each participant took one of the drugs for 6 week and the percent of weight loss was recorded.
}
\examples{
data("weightLoss", package='syllogi')
str(weightLoss)
}
\references{
Fictitious data set
}
\keyword{data}
