% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/federalistPapers.R
\docType{data}
\name{federalistPapers}
\alias{federalistPapers}
\title{Federalist Papers}
\format{
The list has 86 elements, each element is a list with 2 elements.
The paper element is the text of the paper. The meta element is a data frame:
\describe{
  \item{number}{Paper number.}
  \item{author}{Author of the paper.}
  \item{title}{Title of the paper.}
  \item{journal}{Newpaper that published the paper.}
  \item{date}{Date of publication.}
}
}
\usage{
data(federalistPapers)
}
\description{
List of the Federalist Papers
}
\details{
The Project Gutenberg version of the Federalist Papers attributes
paper No. 58 to Madison, but Mosteller and Wallace consider this
paper to have disputed authorship. Thus, this version considers No. 58
authorship to be disputed.

The Project Gutenberg has two slightly different versions of No. 70, both included.
}
\examples{
data("federalistPapers", package='syllogi')
str(federalistPapers)
}
\references{
https://www.gutenberg.org/ebooks/18

Mosteller, F. and D. L. Wallace. Inference and Disputed Authorship: The Federalist. Reading, MA., 1964
}
\keyword{data}
